Flashibi (#17)
Working Flash Attention, in Triton, with bias/mask argument. This branch started with the CUDA implementation, and went through a few revisions, excuse all the commits.

Triton flash attention kernel does not support dropout, so if dropout is set to a nonzero value, we fall back to a PyTorch implementation of attention. (In the future we should check if doing psuedo-dropout by dropping out values from the attention mask before unpadding (or by applying dropout to v) works as well as real dropout. Or we should implement dropout in Triton. But, not knowing Triton, that seems harder.)

This should "just work", however, it is rough around the edges. Code still needs comments and some shape information, and despite main.py implying that writing alibi as an algorithm in your config YAML will cause us to apply ALiBi... we actually just always apply it in the mosaic_bert. Same with GLU. However FLN we do respect. I believe we have to add EMA to main.py for it to work.

Worth testing if amp precision is preferable to bf16. I would expect fp16 to be better in this scenario (I believe fp16 => bf16 usually trades decreased speed for increased stability, but this model is already stable).