make LlamaTokenizer work for tokenizer bakeoff
This adds some HF code that we can delete once we upgrade to a currently unreleased version that supports a certain type of sentencepiece tokenizer.

It also adds sentencepiece as an llm dependency, which is coming soon anyway.

Finally, it imports this code into common/convert_dataset.py and common/text_data.py to make it so AutoTokenizer.from_pretrained can be called with our new custom tokenizers and the Llama tokenizer