Support bot (#378)
Created a demo on how to use MosaicML products, Langchain, and FAISS to create a chatbot!
This PR covers how to get GitHub data, how to convert it into MDS format, how to use that data to fine-tune our model, and how to run retrieval on it to create a chatbot in a gradio app. It also offers different ways to improve the retrieval process to get even better results. Huge thanks to @dakinggg !